[Design] #18 WaitingInfoView 구현
