+++ unifest-ios/Views/Map/OneMapView.swift
@@ -9,6 +9,8 @@ import SwiftUI
 import CoreLocation
 import MapKit
 
+// DetailView에서 '위치 확인하기' 버튼을 누르면 fullscreencover로 뜨는 뷰
+
 @available(iOS 17, *)
 struct OneMapViewiOS17: View {
     @Namespace private var oneMap
@@ -528,6 +530,9 @@ struct OneBoothAnnotation: View {
 }
 
 #Preview {
-    RootView(rootViewModel: RootViewModel())
+    Group {
+        RootView(rootViewModel: RootViewModel())
+        OneBoothAnnotation(boothType: .booth)
+    }
 }
